add Support for REDIS ZSET
This commit adds support for ZSET queues in REDIS with data_type
sortedset.
Batch is supported.

Priority is based on a field of the message containing a number.
If the field is absent or not a number a default priority is then used.

It adds the following options to the configuration:

- priority_field: string, default 'epoch', Priority field to
  use, if field doesn't exist, priority will be priority_default
- priority_default: number, default '-1', Default priority when
  priority field is not found in the event

output {
     redis {
        host => "127.0.0.1"
        port => 6379
        data_type => "sortedset"
	key => "myzip"
	priority_field => "myTimestamp"
   }
}

Signed-off-by: tandara0 <j0.kim@samsung.com>